OLMIS-662: API Data Validation
Add (recommend) a way of performing bean-level validation. Spring provides two options:

1) Injecting an org.springframework.validation.Validator into a controller and manually invoking it.

2) Using JSR 303 / 349 based bean validation, specified via annotations. Optionally, Hibernate-specific validations can be used instead.

The second approach works seamlessly with Spring Data Rest, and is provided by our current Spring Boot configuration out of the box. It’s thus the approach taken and implicitly recommended by this check-in.

As commented within this check-in’s code, an additional and complimentary type of validation is also recommended for use. Specifically, standard JPA annotations should be specified to enforce constraints that belong at the database level. The archetypical example of this is a unique constraint: it’s implication for a specific bean can’t be determined exclusively by looking at the bean (“bean-level validation”), but rather relies on the overall state of all beans within the database (“database level validation”).