Add examples of Validator-based validation
Validation can be performed via a Validator within the context of Spring MVC as well as Spring Data Rest. This check-in demonstrates both approaches.

It should be noted that the Validator-based approach seems rare. The preponderance of tutorials and books instead demonstrate the use of annotations. One book I read even goes so far as to advice against the validator-based approach. (Its reason seemed more a matter of preference than principle. It claims that, just as Spring Annotation-based configuration is cleaner than the XML-based alternative, so too are annotations cleaner than the Validator pattern.)

To elaborate on the above point: even the Spring team seems to prefer annotations over the Validator pattern. This is manifest in their documentation, and corroborated by the fact that bug <https://jira.spring.io/browse/DATAREST-524> has been left open for over a year and marked as a low priority.

A previous commit to this repository demonstrated the use of annotations.