OLMIS-592 Fix i18n example code and document
Add sample code for locales and messages. Remove unused MessageService. Move hello to MessageController, so ServiceNameController does not deal with messages. Change resolver to use cookies, to be more API-oriented. Add Portuguese sample message file.

Document i18n in README.md to point to the one in the Service Template.